+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/spawners/Spawners.java
@@ -102,7 +102,7 @@ public class Spawners extends JsonFeature {
 	public static Integer empoweredMobsAmount = 24;
 	@Config(min = 1)
 	@Label(name = "Empowered.Delay", description = "Spawning Delay (in ticks) when the Spawner is empowered.")
-	public static MinMax empoweredDelay = new MinMax(100, 300);
+	public static MinMax empoweredDelay = new MinMax(150, 300);
 	@Config(min = 0)
 	@Label(name = "Empowered.Experience Reward", description = "When the Spawner stops being empowered, will generate this amount of experience")
 	public static MinMax empoweredExperienceReward = new MinMax(150, 200);