+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Healing potions now heal 3*2^level instead of 2*2^level per level (like pre-1.6.1)
+* Increased spawners delay when empowered and increased experience from broken spawners
 * Fixed Plant growth modifiers not working properly
   * Fixes nether wart not growing and some plants growing slower than they should